Propagated "long" type for step related variables throughout Protomol. Had to add a default for number of steps (InputNumsteps::registerConfiguration(config,1L);) to prevnt the occurance of a bug that cannot find a "check" method in "Value.h/cpp" for LONG type. Put InputNumsteps::registerConfiguration(config); to reproduce un-fixed bug.
git-svn-id: http://svn.code.sf.net/p/protomol/svn/trunk@2086 ce46b6ef-3445-0410-911d-c47028e6f9df

+++ protomol/src/protomol/ProtoMolApp.h
@@ -69,7 +69,7 @@ namespace ProtoMol {
     bool configure(const std::vector<std::string> &args);
     void build();
     void print(std::ostream &stream);
-    bool step(unsigned inc = 0);
+    bool step(long inc = 0);
     void finalize();
   };
 }
